upd Repositories: added findBy methods and refactoring models

+++ src/main/java/ru/ketbiev/hov/model/Account.java
@@ -9,5 +9,5 @@ public class Account {
     private String password;
     private String url;
     private String text;
-    private int spaceId;
+    private long spaceId;
 }
