+++ src/main/java/ru/ketbiev/hov/model/Note.java
@@ -7,5 +7,5 @@ public class Note {
     private long id;
     private String title;
     private String text;
-    private int spaceId;
+    private long spaceId;
 }
